move main.py and getty.py to getty module
