+++ mediachain/getty/ingest.py
